// apps/trip-go/src/types/tour.ts
// types/AreaData.ts
export interface AreaData {
    addr1: string;
    addr2: string;
    areacode: string;
    booktour: string;
    cat1: string;
    cat2: string;
    cat3: string;
    contentid: string;
    contenttypeid: string;
    cpyrhtDivCd: string;
    createdtime: string;
    firstimage: string;
    firstimage2: string;
    mapx: string;
    mapy: string;
    mlevel: string;
    modifiedtime: string;
    sigungucode: string;
    tel: string;
    title: string;
    zipcode: string;
}